category and brand delete bug fixed

+++ app/Http/Controllers/Api/BrandController.php
@@ -94,6 +94,9 @@ class BrandController extends Controller
     private function deleteImage(Brand $brand)
     {
         $image = $brand->image;
+        if (is_null($image)) {
+            return false;
+        }
         File::delete(public_path() . $image->address);
         $image->delete();
     }
